requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
certifi==2024.8.30
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
goodfire==0.2.23
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipysigma==0.24.3
ipython==8.29.0
ipywidgets==8.1.5
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
widgetsnbextension==4.0.13
